mm: Change memory_limit from phys_addr_t to unsigned long long
Patchwork-id: 55326
O-Subject: [RHEL6.4 PATCH 1/2] [powerpc] Change memory_limit from phys_addr_t to unsigned long long
Bugzilla: 674664
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

bz674664
https://bugzilla.redhat.com/show_bug.cgi?id=674664

This is a backport of upstream commit.
In upstream,first_memblock_size is used to track the size of the very
first block in device tree as it represents the RMA on ppc64 server.
In rhel6, the tracking is not imported,change related to first_memblock_size
from upstream is ignored.

commit a84fcd46870113e92523e1ebb9a0ec75f66e03a2
Author: Suzuki Poulose <suzuki@in.ibm.com>
Date:   Tue Aug 21 01:42:33 2012 +0000

    powerpc: Change memory_limit from phys_addr_t to unsigned long long

    There are some device-tree nodes, whose values are of type phys_addr_t.
    The phys_addr_t is variable sized based on the CONFIG_PHSY_T_64BIT.

    Change these to a fixed unsigned long long for consistency.

    This patch does the change only for memory_limit.

    The following is a list of such variables which need the change:

     1) kernel_end, crashk_size - in arch/powerpc/kernel/machine_kexec.c

     2) (struct resource *)crashk_res.start - We could export a local static
        variable from machine_kexec.c.

    Changing the above values might break the kexec-tools. So, I will
    fix kexec-tools first to handle the different sized values and then change
     the above.

    Suggested-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

+++ arch/powerpc/kernel/prom.c
@@ -426,7 +426,7 @@ static int __init early_parse_mem(char *p)
 		return 1;
 
 	memory_limit = PAGE_ALIGN(memparse(p, &p));
-	DBG("memory limit = 0x%llx\n", (unsigned long long)memory_limit);
+	DBG("memory limit = 0x%llx\n", memory_limit);
 
 	return 0;
 }
